+++ hbp_sphinx_theme/__init__.py
@@ -5,7 +5,7 @@ import os
 
 # This file has been adapted from
 # https://github.com/snide/sphinx_rtd_theme/blob/master/sphinx_rtd_theme/__init__.py
-__version__ = '0.3.2'
+__version__ = '0.3.3'
 
 
 def get_html_theme_path():