+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+* added release version info to theme.conf options section
+
 ### 0.3.2 - 2015-08-17
 * hbp-collaboratory-sphinx-theme is available as release zip asset