+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+### 0.3.3 - 2016-01-04
 * added release version info to theme.conf options section
 
 ### 0.3.2 - 2015-08-17